@@ -10,26 +10,23 @@
     "build:prod": "npm run build && npm run prerender",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "dev:ssr": "ng run schlosswochen-ch:serve-ssr",
-    "serve:ssr": "node dist/schlosswochen-ch/server/main.js",
-    "build:ssr": "ng build && ng run schlosswochen-ch:server",
-    "prerender": "ng run schlosswochen-ch:prerender"
+    "serve:ssr:schlosswochen-ch": "node dist/schlosswochen-ch/server/server.mjs"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.5",
+    "@angular/animations": "^18.2.13",
     "@angular/cdk": "^17.0.2",
-    "@angular/common": "^17.0.5",
-    "@angular/compiler": "^17.0.5",
-    "@angular/core": "^17.0.5",
-    "@angular/forms": "^17.0.5",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
     "@angular/google-maps": "^17.3.10",
     "@angular/material": "^17.0.2",
-    "@angular/platform-browser": "^17.0.5",
-    "@angular/platform-browser-dynamic": "^17.0.5",
-    "@angular/platform-server": "^17.0.5",
-    "@angular/router": "^17.0.5",
-    "@angular/ssr": "^17.0.5",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/platform-server": "^18.2.13",
+    "@angular/router": "^18.2.13",
+    "@angular/ssr": "^18.2.20",
     "@netlify/functions": "^1.4.0",
     "axios": "^0.27.2",
     "express": "^4.18.2",
@@ -38,13 +35,13 @@
     "rxjs": "~7.4.0",
     "swiper": "^11.0.5",
     "tslib": "^2.3.0",
-    "zone.js": "~0.14.2"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.5",
-    "@angular/cli": "^17.0.5",
-    "@angular/compiler-cli": "^17.0.5",
-    "@angular/localize": "^17.0.5",
+    "@angular-devkit/build-angular": "^18.2.20",
+    "@angular/cli": "^18.2.20",
+    "@angular/compiler-cli": "^18.2.13",
+    "@angular/localize": "^18.2.13",
     "@tailwindcss/forms": "^0.5.7",
     "@tailwindcss/typography": "^0.5.10",
     "@types/express": "^4.17.17",
@@ -63,6 +60,6 @@
     "prettier": "2.6.2",
     "tailwindcss": "^3.3.5",
     "ts-node": "^10.8.0",
-    "typescript": "~5.2.2"
+    "typescript": "~5.5.4"
   }
 }
